Split mtls checker
